+++ src/components/child/error.tsx
@@ -1,0 +1,4 @@
+// error.tsx
+const ErrorMessage = ({ message }: { message: string }) => <p>오류가 발생했습니다: {message}</p>;
+
+export default ErrorMessage;
