+++ src/components/child/loading.tsx
@@ -1,0 +1,3 @@
+const Loading = () => <p>로딩 중...</p>;
+
+export default Loading;
